♻️  zb: Refactor e2e test
