+++ zbus/tests/iface_and_proxy/mod.rs
@@ -1,0 +1,4 @@
+pub mod client;
+pub mod helpers;
+pub mod iface;
+pub mod types;
